requirements-lock.txt
# Pinned dependency versions for reproducible development installs.
# Generated with: pip freeze
# Date: 2026-02-22
#
# This file is for DEVELOPMENT REPRODUCIBILITY only.
# Production installs should use: pip install rain-assistant
# (which uses the >= constraints in pyproject.toml)
#
# To use:  pip install -r requirements-lock.txt
# To regenerate:  pip freeze > requirements-lock.txt
#
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altgraph==0.17.5
annotated-types==0.7.0
anthropic==0.79.0
anyio==4.9.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
auth0-python==4.10.0
av==16.1.0
backoff==2.2.1
backports.asyncio.runner==1.2.0
backports.tarfile==1.2.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chromadb==1.0.12
claude-agent-sdk==0.1.36
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
crewai==0.130.0
cryptography==45.0.4
ctranslate2==4.7.1
decorator==5.2.1
deep-translator==1.11.4
deta==1.2.0
distro==1.9.0
docstring_parser==0.16
docutils==0.22.4
durationpy==0.10
edge-tts==7.2.7
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.9
faster-whisper==1.2.1
filelock==3.16.1
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2024.10.0
google-auth==2.38.0
googleapis-common-protos==1.70.0
grpcio==1.73.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.33.0
humanfriendly==10.0
id==1.6.1
idna==3.10
ijson==3.4.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
instructor==1.8.3
ipython==8.37.0
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
json-stream==2.3.3
json-stream-rs-tokenizer==0.4.29
json5==0.12.0
json_repair==0.46.2
jsonpickle==4.1.1
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keyring==25.7.0
kubernetes==33.1.0
litellm==1.72.0
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mcp==1.26.0
mdurl==0.1.2
mmh3==5.1.0
more-itertools==10.8.0
MouseInfo==0.1.3
mpmath==1.3.0
mss==10.1.0
multidict==6.4.4
networkx==3.4.2
nh3==0.3.3
numpy==2.1.3
oauthlib==3.2.2
ollama==0.6.1
onnxruntime==1.22.0
openai==1.88.0
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
orjson==3.10.18
overrides==7.7.0
packaging==25.0
parso==0.8.4
pdfminer.six==20250506
pdfplumber==0.11.7
pefile==2024.8.26
pillow==11.0.0
pluggy==1.6.0
posthog==5.0.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.5
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycloudflared==0.3.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.12.0
pydantic_core==2.33.2
PyGetWindow==0.0.9
Pygments==2.19.1
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.0
PyJWT==2.10.1
PyMsgBox==2.0.1
pypdfium2==4.30.1
pyperclip==1.11.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
python-multipart==0.0.22
pytweening==1.2.0
pyvis==0.3.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.2
readme_renderer==44.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
rpds-py==0.25.1
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.5
soupsieve==2.7
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.45.3
sympy==1.13.3
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
tomli_w==1.2.0
torch==2.4.1
torch-directml==0.2.5.dev240914
torchvision==0.19.1
tqdm==4.67.1
traitlets==5.14.3
twine==6.2.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.4.0
uv==0.7.13
uvicorn==0.34.3
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
